requirements.txt
streamlit==1.24.0
opencv-python
pillow
numpy
psycopg2